Move ScalingAlongCardinalAxes to specialised implementation.
While the previous solution would work for all non-specialised matrices, it would generate a new function for each parameter combination that's a different type. Ideally they would all be of the same type as each element contained within the matrix, something like ``T... Args`` but that's not valid. Using concepts to require each parameter in the pack be the same as T will just add unnecesary vebosity.