Switch to deducing this
Not sure if this is better or not.
Can avoid having to pass in the derived class and the subsequent casting to it, but it means there's a lot more repetitive typing unless auto is used.

I'm not the biggest fan of auto because it can make things less obvious, but as the only times it's used here is on operators where you won't care to check the result, and on IdentityMatrix which is obvious, I've used it.