Update Vector2/PolarCoordinates convert methods
I felt that handling the conversions to the other class in each was more intutitve than having it on only one class, and that methods were self documenting compared to using constructors.

This did result in a circular dependency had to be solved by forward declaring each object for the declaration, and then including the header for the implementation. I actually realised that doing all of this esclusively within the header file was possible until now.